+++ test/squint/string_test.cljs
@@ -1,15 +1,15 @@
 (ns squint.string-test
   (:require
-   [clojure.test :as t :refer [deftest]]
+   [clojure.test]
    [squint.compiler :as compiler]
-   [squint.test-utils :refer [eq]]
    #_:clj-kondo/ignore
    ["fs" :as fs]
    #_:clj-kondo/ignore
    ["path" :as path]
-   ;; required at least by the `squint.eval-macro/evalll` macro
+   ;; required by the `squint.eval-macro/deftest-eval` macro
+   #_:clj-kondo/ignore
    ["url" :as url])
-  (:require-macros [squint.eval-macro :refer [evalll]]))
+  (:require-macros [squint.eval-macro :refer [deftest-eval]]))
 
 (defn compile! [str-or-expr]
   (let [s (if (string? str-or-expr)
@@ -19,84 +19,70 @@
 
 (def dyn-import (js/eval "(x) => import(x)"))
 
-(deftest ^:async blank?-test
-  (evalll true
-          '(do (ns foo (:require [squint.string :as str]))
-               (def result (str/blank? "")))))
+(deftest-eval blank?-test
+  (do (ns foo (:require [squint.string :as str]
+                        [cljs.test :refer [is]]))
+      (is (str/blank? ""))))
 
-(deftest ^:async join-test
-  (evalll "0--1--2--3--4--5--6--7--8--9"
-          '(do (ns foo (:require [squint.string :as str]))
-               (def result (str/join "--" (range 10))))))
+(deftest-eval join-test
+  (do (ns foo (:require [squint.string :as str]
+                        [cljs.test :refer [is]]))
+      (is (= "0--1--2--3--4--5--6--7--8--9" (str/join "--" (range 10))))))
 
-(deftest ^:async replace-test
-  (evalll "yyxxyyxx"
-          '(do (ns foo (:require [squint.string :as str]))
-               (def result (str/replace "--xx--xx" "--" "yy")))))
+(deftest-eval replace-test
+  (do (ns foo (:require [squint.string :as str]
+                        [cljs.test :refer [is]]))
+      (is (= "yyxxyyxx" (str/replace "--xx--xx" "--" "yy")))))
 
-(deftest ^:async split-test
-  (evalll (eq ["foo" "bar"])
-          '(do (ns foo (:require [squint.string :as str]))
-               (def result (str/split "foo\nbar\n\n" #"\n")))))
+(deftest-eval replace-first-test
+  (do (ns foo (:require [squint.string :as str]
+                        [cljs.test :refer [is]]))
+      (is (= "yyxx--xx" (str/replace-first "--xx--xx" "--" "yy")))
+      (is (= "yyxx--xx" (str/replace-first "--xx--xx" #"--" "yy")))
+      (is (= "a[b]c" (str/replace-first "abc" #"(b)" "[$1]")))
+      (is (= "a$xc" (str/replace-first "abc" "b" "$x")))
+      (is (= "abc" (str/replace-first "abc" #"x" "y")))))
 
-(deftest ^:async split-limit-test
-  ;; positive limit caps the number of splits and keeps the remainder,
-  ;; matching Clojure (not JS truncation)
-  (evalll (eq ["a" "b-c-d"])
-          '(do (ns foo (:require [squint.string :as str]))
-               (def result (str/split "a-b-c-d" #"-" 2))))
-  (evalll (eq ["a" "b" "c-d"])
-          '(do (ns foo (:require [squint.string :as str]))
-               (def result (str/split "a-b-c-d" #"-" 3))))
-  ;; limit 0 / unset discards trailing empties; negative keeps them
-  (evalll (eq ["a" "b"])
-          '(do (ns foo (:require [squint.string :as str]))
-               (def result (str/split "a-b-" #"-" 0))))
-  (evalll (eq ["a" "b" ""])
-          '(do (ns foo (:require [squint.string :as str]))
-               (def result (str/split "a-b-" #"-" -1)))))
+(deftest-eval split-test
+  (do (ns foo (:require [squint.string :as str]
+                        [cljs.test :refer [is]]))
+      (is (= ["foo" "bar"] (str/split "foo\nbar\n\n" #"\n")))))
 
-(deftest ^:async split-lines-test
-  (evalll (eq ["foo" "bar"])
-          '(do (ns foo (:require [squint.string :as str]))
-               (def result (str/split-lines "foo\nbar\n\n")))))
+(deftest-eval split-limit-test
+  (do (ns foo (:require [squint.string :as str]
+                        [cljs.test :refer [is]]))
+      ;; positive limit caps the number of splits and keeps the remainder,
+      ;; matching Clojure (not JS truncation)
+      (is (= ["a" "b-c-d"] (str/split "a-b-c-d" #"-" 2)))
+      (is (= ["a" "b" "c-d"] (str/split "a-b-c-d" #"-" 3)))
+      ;; limit 0 / unset discards trailing empties; negative keeps them
+      (is (= ["a" "b"] (str/split "a-b-" #"-" 0)))
+      (is (= ["a" "b" ""] (str/split "a-b-" #"-" -1)))))
 
-(deftest ^:async lower-case-test
-  (evalll "foobar"
-          '(do (ns foo (:require [squint.string :as str]))
-               (def result (str/lower-case "FooBar")))))
+(deftest-eval split-lines-test
+  (do (ns foo (:require [squint.string :as str]
+                        [cljs.test :refer [is]]))
+      (is (= ["foo" "bar"] (str/split-lines "foo\nbar\n\n")))))
 
-(deftest ^:async upper-case-test
-  (evalll "FOOBAR"
-          '(do (ns foo (:require [squint.string :as str]))
-               (def result (str/upper-case "FooBar")))))
+(deftest-eval lower-case-test
+  (do (ns foo (:require [squint.string :as str]
+                        [cljs.test :refer [is]]))
+      (is (= "foobar" (str/lower-case "FooBar")))))
 
-(deftest ^:async capitalize-test
-  (evalll (eq ["" "F" "Foobar"])
-          '(do (ns foo (:require [squint.string :as str]))
-               (def result [(str/capitalize "")
-                            (str/capitalize "f")
-                            (str/capitalize "FooBar")]))))
+(deftest-eval upper-case-test
+  (do (ns foo (:require [squint.string :as str]
+                        [cljs.test :refer [is]]))
+      (is (= "FOOBAR" (str/upper-case "FooBar")))))
 
-(deftest ^:async includes-test
-  (evalll (eq [true])
-          '(do (ns foo (:require [squint.string :as str]))
-               (def result [(str/includes? "foo" "o")]))))
+(deftest-eval capitalize-test
+  (do (ns foo (:require [squint.string :as str]
+                        [cljs.test :refer [is]]))
+      (is (= ["" "F" "Foobar"]
+             [(str/capitalize "")
+              (str/capitalize "f")
+              (str/capitalize "FooBar")]))))
 
-;; (deftest ^:async string-conflict-test
-;;   (evalll (fn [res]
-;;             (eq ["foo","bar"] res))
-;;           '(do (ns foo (:require [squint.string :as str]))
-;;                (defn split [x] (str x)) (def result (str/split "foo,bar" ",")))))
-
-;; (deftest ^:async split-test-string
-;;   (evalll (fn [res]
-;;             (eq ["foo","bar","baz"] res))
-;;           '(do (ns foo (:require [squint.string :as str]))
-;;                (def result (str/split "foo--bar--baz" "--")))))
-
-;; (deftest ^:async split-test-regex
-;;   (evalll (fn [res]
-;;             (eq ["foo","bar","baz"] res))
-;;           '(do (ns foo (:require [squint.string :as str]))
-;;                (def result (str/split "fooxbarybaz" #"[xy]")))))
+(deftest-eval includes-test
+  (do (ns foo (:require [squint.string :as str]
+                        [cljs.test :refer [is]]))
+      (is (str/includes? "foo" "o"))))